+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.9.2 (unreleased)
 
+  - Support shell expansions in shared folder guest paths again. [GH-656]
   - Fix `forward_agent` not working when outside of blocks. [GH-651]
   - Fix issue causing custom guest implementations to not load properly.
   - Filter clear screen character out of output on SSH.